Resolve more circular dependencies in rascal.utils
