+++ bindings/rascal/utils/__init__.py
@@ -7,27 +7,14 @@ from .io import (
     dump_obj,
     load_obj,
 )
+from .misc import is_notebook
+
+# Warning potential dependency loop: FPS imports models, which imports KRR,
+# which imports this file again
 from .fps import fps, FPSFilter
+
+# function to redirect c++'s standard output to python's one
 from ..lib._rascal.utils import ostream_redirect
-from ..lib import utils
 from copy import deepcopy
 from .cur import CURFilter
 from .scorer import get_score, print_score
-
-# function to redirect c++'s standard output to python's one
-ostream_redirect = utils.__dict__["ostream_redirect"]
-
-
-def is_notebook():
-    from IPython import get_ipython
-
-    try:
-        shell = get_ipython().__class__.__name__
-        if shell == "ZMQInteractiveShell":
-            return True  # Jupyter notebook or qtconsole
-        elif shell == "TerminalInteractiveShell":
-            return False  # Terminal running IPython
-        else:
-            return False  # Other type (?)
-    except NameError:
-        return False  # Probably standard Python interpreter
