+++ bindings/rascal/utils/misc.py
@@ -1,0 +1,17 @@
+"""Miscellaneous useful utilities"""
+
+
+def is_notebook():
+    """Is this being run inside an IPython Notebook?"""
+    from IPython import get_ipython
+
+    try:
+        shell = get_ipython().__class__.__name__
+        if shell == "ZMQInteractiveShell":
+            return True  # Jupyter notebook or qtconsole
+        elif shell == "TerminalInteractiveShell":
+            return False  # Terminal running IPython
+        else:
+            return False  # Other type (?)
+    except NameError:
+        return False  # Probably standard Python interpreter
